LPS-66932 Move KB data handlers to the proper package
